move standard events
