+++ frontend/app/src/entities/events/ui/global-event.tsx
@@ -6,9 +6,9 @@ import { Icon } from "@iconify-icon/react";
 import { format } from "date-fns";
 import { BRANCH_EVENTS, GROUP_EVENTS, STANDARD_EVENTS } from "../constants";
 import { BranchEvent } from "./branch-events/global-branch-event";
-import { StandardEvent } from "./global-standard-event";
 import { GroupEvent } from "./group-events/global-group-event";
 import { NodeEvent } from "./node-events/global-node-event";
+import { StandardEvent } from "./standard-events/global-standard-event";
 
 const GlobalEventDisplay = ({ __typename, ...props }: EventType) => {
   if ("attributes" in props) {
