+++ frontend/app/src/entities/events/ui/standard-events/global-standard-event.tsx
