+++ frontend/app/src/entities/events/ui/standard-events/standard-event.tsx
